pipeline/requirements.txt
PyYAML
edgartools
pydantic
# azure-core
# azure-ai-documentintelligence
# azure-ai-formrecognizer
python-dotenv
